workers: clarify the default port number and Unix sockets support
The default conf/worker-proxy.inc config example omits the port number
for "localhost", it was not obvious was port number will be used.

Add an example for a Unix socket. This appears to work as expected, but
does not support priorities at the moment. See also the comment in
src/libutil/addr.c, function rspamd_parse_host_port_priority.