Added rspamd_logo_white_bg.png to update on readme
Currently on rspamd/rspamd the readme file has the logo that is not clear when dark mode is turned on. Adding this image and updating the url on rspamd/rspamd repository will solve this issue. There is another way as well, to switch between logos according to light or dark mode, but then we can not attach a link to it, or change its size with HTML. So this is the current best workaround,